# swagger/swagger.yml
swagger: '2.0'
info:
  title: QbVIP backend
  description: | 
    graphiql:   http://localhost:19931/graphiql
  version: 1.0.0
paths:

  # 用户部分 start
  /v1/users:
    get:
      tags:
        - user
      description: 获取所有用户
      parameters:
        - in: header
          name: token
          schema:
            type: string
            format: uuid
          required: true
      responses:
        '200':
          description: ''
        '400':
          description: no enough input
  /v1/users/{id}:
    get:
      tags:
        - user
      parameters:
        - in: header
          name: token
          schema:
            type: string
            format: uuid
          required: true
        - in: path
          name: id
          description: 用户ID
      description: 获取某用户的信息
      responses:
        '200':
          description: ''
        '400':
          description: no enough input
    delete:
      tags:
        - user
      parameters:
        - in: header
          name: token
          schema:
            type: string
            format: uuid
          required: true
        - in: path
          name: id
          description: 用户ID
      description: 删除用户
      responses:
        '200':
          description: ''
        '400':
          description: no enough input
  /v1/users/login:
    post:
      tags:
        - user
      description: 登录系统
      parameters:        
        - in: body
          name: body
          description: 用户信息
          schema:
            $ref: '#/definitions/userLogin'   
      responses:
        '200':
          description: ''
        '400':
          description: no enough input
  /v1/users/logout:
    delete:
      tags:
        - user
      description: 登出系统
      parameters:        
        - in: header
          name: token
          schema:
            type: string
            format: uuid
          required: true
      responses:
        '200':
          description: ''
        '400':
          description: no enough input
  /v1/users/loginname/{loginName}:
    get:
      tags:
        - user
      parameters:
        - in: header
          name: token
          schema:
            type: string
            format: uuid
          required: true
        - in: path
          name: loginName
          description: 用户登录名
      description: 根据用户登录名获取用户信息
      responses:
        '200':
          description: ''
        '400':
          description: no enough input
  /v1/users/registry:
    post:
      tags:
        - user
      description: 用户注册
      parameters:
        - in: body
          name: body
          description: 用户信息
          schema:
            $ref: '#/definitions/userinfo'
      responses:
        '200':
          description: ''
        '400':
          description: no enough input
  # 用户部分 end  

  # 网页收藏夹 start
  /v1/favorites/folder:
    get:
      tags:
        - favorites_folder
      description: 获取所有的收藏夹文件夹
      parameters:
        - in: header
          name: token
          schema:
            type: string
            format: uuid
          required: true
      responses:
        '200':
          description: ''
        '400':
          description: no enough input
    post:
      tags:
        - favorites_folder
      description: 添加收藏夹
      parameters:
        - in: header
          name: token
          schema:
            type: string
            format: uuid
          required: true
        - in: body
          name: body
          description: 收藏夹信息
          schema:
            $ref: '#/definitions/favorites_folder'
      responses:
        '200':
          description: ''
        '400':
          description: no enough input
  /v1/favorites/folder/{folderID}:    
    put:
      tags:
        - favorites_folder
      description: 添加收藏夹
      parameters:
        - in: header
          name: token
          schema:
            type: string
            format: uuid
          required: true
        - in: path
          name: folderID
          description: 收藏夹ID
        - in: body
          name: body
          description: 收藏夹信息
          schema:
            $ref: '#/definitions/favorites_folder'
      responses:
        '200':
          description: ''
        '400':
          description: no enough input
    delete:
      tags:
        - favorites_folder
      description: 删除收藏夹文件夹
      parameters:
        - in: header
          name: token
          schema:
            type: string
            format: uuid
          required: true
        - in: path
          name: folderID
          description: 收藏夹目录ID
      responses:
        '200':
          description: ''
        '400':
          description: no enough input

  # 收藏夹站点 start
  /v1/favorites/website:
    get:
      tags:
        - favorites
      description: 获取所有的收藏站点
      parameters:
        - in: header
          name: token
          schema:
            type: string
            format: uuid
          required: true
      responses:
        '200':
          description: ''
        '400':
          description: no enough input
    post:
      tags:
        - favorites
      description: 添加收藏夹
      parameters:
        - in: header
          name: token
          schema:
            type: string
            format: uuid
          required: true
        - in: body
          name: body
          description: 收藏夹信息
          schema:
            $ref: '#/definitions/favorites'
      responses:
        '200':
          description: ''
        '400':
          description: no enough input               
  /v1/favorites/website/{wsID}:
    put:
      tags:
        - favorites
      description: 编辑收藏夹
      parameters:
        - in: header
          name: token
          schema:
            type: string
            format: uuid
          required: true
        - in: path
          name: wsID
          description: 站点ID
        - in: body
          name: body
          description: 站点内容
          schema:
            $ref: '#/definitions/favorites'
      responses:
        '200':
          description: ''
        '400':
          description: no enough input
    delete:
      tags:
        - favorites
      description: 删除收藏夹
      parameters:
        - in: header
          name: token
          schema:
            type: string
            format: uuid
          required: true
        - in: path
          name: wsID
          description: 站点ID     
      responses:
        '200':
          description: ''
        '400':
          description: no enough input
  /v1/favorites/website/import:
    post:
      tags:
        - favorites
      description: 导入收藏夹文件
      parameters:
        - in: header
          name: token
          schema:
            type: string
            format: uuid
          required: true        
        - in: body
          name: body
          description: 站点内容
          schema:
            $ref: '#/definitions/favorites_import'
      responses:
        '200':
          description: ''
        '400':
          description: no enough input
          # 收藏夹站点 end

  /v1/accounts:
    get:
      tags:
        - account
      description: 获取所有的账号
      parameters:
        - in: header
          name: token
          schema:
            type: string
            format: uuid
          required: true        
      responses:
        '200':
          description: ''
        '400':
          description: no enough input
    post:
      tags:
        - account
      description: 创建账号
      parameters:
        - in: header
          name: token
          schema:
            type: string
            format: uuid
          required: true    
        - in: body
          name: body
          description: 账号内容
          schema:
            $ref: '#/definitions/account'    
      responses:
        '200':
          description: ''
        '400':
          description: no enough input
    put:
      tags:
        - account
      description: 编辑账号
      parameters:
        - in: header
          name: token
          schema:
            type: string
            format: uuid
          required: true        
        - in: body
          name: body
          description: 账号内容
          schema:
            $ref: '#/definitions/account'
      responses:
        '200':
          description: ''
        '400':
          description: no enough input
  /v1/accounts/{accountUUID}:    
    delete:
      tags:
        - account
      description: 删除账号
      parameters:
        - in: header
          name: token
          schema:
            type: string
            format: uuid
          required: true        
        - in: path
          name: accountUUID
          description: 账号UUID
      responses:
        '200':
          description: ''
        '400':
          description: no enough input
      
  /v1/common/upload:    
    post:      
      summary: Uploads a file.
      consumes:
        - multipart/form-data
      parameters:
        - in: header
          name: token
          schema:
            type: string
            format: uuid
          required: true  
        - in: formData
          name: upfile
          type: file
          description: The file to upload.

definitions:
  userLogin:
    type: object
    properties:
      loginName:
        type: string 
      password:
        type: string
  userinfo:
    type: object
    properties:
      loginName:
        type: string
      nikeName:
        type: string
      email:
        type: string
      password:
        type: string  
  favorites:
    type: object
    properties:      
      name:
        type: string 
      desc:
        type: string 
      label:
        type: array
      iconURL:
        type: string
      url:
        type: string
      folderID:
        type: string

  favorites_folder:
    type: object
    properties:
      parentID:
        type: string
      name:
        type: string
      desc:
        type: string
      order:
        type: number
  account:
    type: object
    properties:      
      url:
        type: string
      type:
        type: string
      accountName:
        type: string
      password:
        type: string
      labels:
        type: array
      common:
        type: string
  favorites_import:
      type: Object
      properties:
        filename:
          type: string
        folderID:
          type: string

tags:
  - name: user
    description: 用户相关API
  - name: favorites_folder
    description: 收藏夹目录
  - name: favorites
    description: 收藏站点
  - name: account
    description: 账号管理